+++ tests/unit-tests/SettingsTest.php
@@ -10,6 +10,10 @@ class SettingsTest extends TestCase {
 
 	public function setUp(): void {
 		parent::setUp();
+		$kpPluginFeatures = Mockery::mock( 'overload:Krokedil\Klarna\PluginFeatures' );
+		$kpPluginFeatures->shouldReceive('is_available')->with( 'klarna-express-checkout:1-step' )->andReturn( false );
+		$kpPluginFeatures->shouldReceive('is_available')->with( 'klarna-express-checkout:2-step' )->andReturn( true );
+		$kpPluginFeatures->shouldReceive('get_acquiring_partner_key')->with()->andReturn( false );
 		WP_Mock::userFunction( 'get_option' )->with( 'kec_webhook', array() )->andReturn( array() );
 		WP_Mock::userFunction( 'get_option' )->with( 'kec_signing_key', array() )->andReturn( array() );
 		WP_Mock::userFunction( 'get_option' )->with( 'kp_unavailable_feature_ids', array() )->andReturn( array() );